Key design overhaul, phase 1
It turns out that you can safely convert an ed25519 key to a curve25519
key, so I am changing from separate signing and encryption keys to a
single seed that generates both keys.  This is a major overhaul, so it
is being done in phases.  This first phase includes modified versions of
nacl.js and nacl-fast.js that export the key converstion function (which
was not included in the original code).